refactor: :truck: remove extraneous router subdirs
